Seller  : restrict access to  product  based on company

+++ src/main/java/org/surja/digital_commerce_backend/service/SellerService.java
@@ -3,6 +3,7 @@ package org.surja.digital_commerce_backend.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.surja.digital_commerce_backend.dto.CreateResponseDTO;
@@ -11,6 +12,7 @@ import org.surja.digital_commerce_backend.dto.ResponseDTO;
 import org.surja.digital_commerce_backend.entity.Category;
 import org.surja.digital_commerce_backend.entity.Company;
 import org.surja.digital_commerce_backend.entity.Product;
+import org.surja.digital_commerce_backend.entity.User;
 import org.surja.digital_commerce_backend.exception.NotFoundException;
 import org.surja.digital_commerce_backend.repo.CategoryRepo;
 import org.surja.digital_commerce_backend.repo.CompanyRepo;
@@ -30,6 +32,10 @@ public class SellerService {
 
     @Transactional
     public CreateResponseDTO createProduct(ProductDTO productDTO){
+
+        User user =(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+
         Product product = new Product();
         product.setName(productDTO.getName());
         product.setDescription(productDTO.getDescription());
@@ -37,24 +43,24 @@ public class SellerService {
         product.setActive(true);
         product.setStocks(productDTO.getStocks());
         product.setImageUrl(productDTO.getImageUrl());
-
-        Company company = companyRepo.findById(productDTO.getCompanyId()).get();
-        product.setCompany(company);
+        product.setCompany(user.getCompany());
 
         Category category = categoryRepo.findById(productDTO.getCategoryId()).get();
         product.setCategory(category);
 
         productRepo.save(product);
 
-        CreateResponseDTO responseDTO = new CreateResponseDTO();
-        responseDTO.setId(product.getId());
-        responseDTO.setMessage("product added succssfully");
 
-        return  responseDTO;
+        return  CreateResponseDTO.builder()
+                .id(product.getId())
+                .errorCode("success - 123")
+                .message("product added succssfully")
+                .build();
     }
 
     public List<ProductDTO> getAllProduct(){
-        List<Product> productList = productRepo.findAll();
+        User user =(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        List<Product> productList = productRepo.findByCompany(user.getCompany());
         List<ProductDTO> result = new ArrayList<>();
         for(Product product : productList){
             ProductDTO productDTO = ProductDTO.buildProductDTOFromProduct(product);
@@ -65,8 +71,13 @@ public class SellerService {
 
     @Transactional
     public ResponseDTO updateProduct(Long id, ProductDTO productDTO) throws NotFoundException {
-        Product product = productRepo.findById(id).orElseThrow(
-                ()-> new NotFoundException("Product Doesn't exist with id : "+id));
+        Product product = productRepo.findById(id).get();
+
+        User user =(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(product == null || ! product.getCompany().equals(user.getCompany())){
+            throw new NotFoundException(" Company having id : "+user.getCompany().getId()+" does not have the product");
+
+        }
 
         product.setName(productDTO.getName());
         product.setPrice(productDTO.getPrice());
@@ -74,7 +85,7 @@ public class SellerService {
         product.setStocks(productDTO.getStocks());
         product.setImageUrl(productDTO.getImageUrl());
         product.setCategory(categoryRepo.findById(productDTO.getCategoryId()).get());
-        product.setCompany(companyRepo.findById(productDTO.getCompanyId()).get());
+        product.setCompany(user.getCompany());
 
 
         ResponseDTO responseDTO = ResponseDTO.builder()
@@ -91,6 +102,11 @@ public class SellerService {
     public ResponseDTO deleteProduct(Long id) throws NotFoundException {
         Product product = productRepo.findById(id).orElseThrow(
                 ()-> new NotFoundException("Product doesn't exists with id : "+id));
+        User user =(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(product == null || ! product.getCompany().equals(user.getCompany())){
+            throw new NotFoundException(" Company having id : "+user.getCompany().getId()+" does not have the product");
+
+        }
 
         productRepo.deleteById(id);
         ResponseDTO responseDTO = new ResponseDTO();
